Use float for flex factor

+++ widget/flex.go
@@ -45,7 +45,7 @@ func (f *Flex) CreateElement() Element {
 }
 
 type Flexible struct {
-	Flex  int
+	Flex  float32
 	Fit   render.FlexFit
 	Child Widget
 }
